pyproject.toml
[project]
name = "klinedataprovider"
version = "0.1.0"
description = ""
authors = [
    {name = "wikimodels",email = "young.wiki.models@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "ccxt (==4.3.50)",
    "pandas (==2.2.2)",
    "pandas-ta-classic (==0.3.14b1)",
    "psycopg2-binary (==2.9.9)",
    "requests (==2.32.3)",
    "hurst (==0.0.5)",
    "scipy (==1.13.1)",
    "tenacity (==8.3.0)",
    "numpy (==1.26.4)",
    "numba (==0.61.2)",
    "fastapi (==0.111.0)",
    "uvicorn[standard] (==0.29.0)",
    "tqdm (==4.66.4)",
    "pymongo (==4.8.0)",
    "aiohttp (>=3.13.0,<4.0.0)",
    "polars (>=1.34.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
